refactor: Refactoring/Reorganizing the api/auth/ source
- Reorganized the api/auth source to be in type specific subdirectories.

  This gives us:
    - api/vault/lightspeed/
    - api/vault/hashicorp_vault/

  with type specific auth.py, serializer.py and view.py instead of the
  shared view and serializer files or api/vault/auth_* named files.
