+++ quipucords/api/auth/hashicorp_vault/auth.py
