+++ quipucords/api/views.py
@@ -3,8 +3,8 @@
 # ruff: noqa: F401
 
 from api.aggregate_report.view import aggregate_report
-from api.auth.view import (
-    HashiCorpVaultViewSet,
+from api.auth.hashicorp_vault.view import HashiCorpVaultViewSet
+from api.auth.lightspeed.view import (
     lightspeed_auth_login,
     lightspeed_auth_logout,
     lightspeed_auth_status,
